refactored to move components to shared directory
