+++ src/applications/vaos/appointment-list/components/AppointmentsPage/AppointmentColumnLayout.jsx
@@ -3,8 +3,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { useSelector } from 'react-redux';
 import moment from 'moment';
-import AppointmentColumn from './AppointmentColumn';
-import AppointmentRow from './AppointmentRow';
+import AppointmentRow from '../../../components/AppointmentRow';
 import {
   selectAppointmentLocality,
   selectIsCanceled,
@@ -18,6 +17,7 @@ import {
   selectModalityAriaText,
   selectIsCommunityCare,
 } from '../../redux/selectors';
+import AppointmentColumn from '../../../components/AppointmentColumn';
 
 export default function AppointmentColumnLayout({
   data,
