+++ src/applications/vaos/appointment-list/components/AppointmentsPage/RequestAppointmentLayout.jsx
@@ -2,10 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { useSelector } from 'react-redux';
-import AppointmentFlexGrid from './AppointmentFlexGrid';
-import ListItem from './ListItem';
-import AppointmentRow from './AppointmentRow';
-import AppointmentColumn from './AppointmentColumn';
+import AppointmentFlexGrid from '../../../components/AppointmentFlexGrid';
+import ListItem from '../../../components/ListItem';
+import AppointmentRow from '../../../components/AppointmentRow';
 import {
   selectAppointmentLocality,
   selectIsCanceled,
@@ -18,6 +17,7 @@ import {
   selectFeatureBreadcrumbUrlUpdate,
   selectFeatureCCDirectScheduling,
 } from '../../../redux/selectors';
+import AppointmentColumn from '../../../components/AppointmentColumn';
 
 export default function RequestAppointmentLayout({ appointment, index }) {
   const appointmentLocality = useSelector(() =>
