+++ src/applications/vaos/referral-appointments/components/PendingReferralCard.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { format, parseISO } from 'date-fns';
 import PropTypes from 'prop-types';
-import AppointmentFlexGrid from '../../appointment-list/components/AppointmentsPage/AppointmentFlexGrid';
-import ListItem from '../../appointment-list/components/AppointmentsPage/ListItem';
-import AppointmentRow from '../../appointment-list/components/AppointmentsPage/AppointmentRow';
-import AppointmentColumn from '../../appointment-list/components/AppointmentsPage/AppointmentColumn';
+import AppointmentFlexGrid from '../../components/AppointmentFlexGrid';
+import ListItem from '../../components/ListItem';
+import AppointmentRow from '../../components/AppointmentRow';
+import AppointmentColumn from '../../components/AppointmentColumn';
 
 const PendingReferralCard = ({ referral, handleClick, index }) => {
   const first = index === 0;
